source/backend/requirements.txt
fastapi>=0.68.0,<0.69.0
fastapi-csrf-protect==0.2.1
uvicorn>=0.15.0,<0.16.0
pydantic[email,dotenv]>=1.2.0,<2.0.0
SQLAlchemy==1.3.18
SQLAlchemy-Utils==0.36.8
psycopg2==2.8.6
alembic==1.4.2
starlette-csrf==1.4.4
PyJWT==2.3.0
passlib==1.7.4
pytest==7.4.3
requests==2.24.0